+++ source/Internal/Common/FormConfiguration/FormFieldsConfigurationDataProviderInterface.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\EshopCommunity\Internal\Common\FormConfiguration;
+
+/**
+ * @internal
+ */
+interface FormFieldsConfigurationDataProviderInterface
+{
+    /**
+     * @return array
+     */
+    public function getFormFieldsConfiguration();
+}